+++ src/mockup/standalone.py
@@ -11,10 +11,10 @@ import os
 from pathlib import Path
 import sys
 
+from mockup.aeon2_converter import Aeon2Converter
+from mvclib.user_interface.ui_tk import UiTk
 from mvclib.view.set_icon_tk import set_icon
 from mvclib.view.ui import Ui
-from mvclib.user_interface.ui_tk import UiTk
-from nvaeon2lib.aeon2_converter import Aeon2Converter
 from nvlib.configuration.configuration import Configuration
 from nvlib.nv_locale import _