+++ src/nv_aeon2.py
@@ -16,19 +16,13 @@ but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 """
-from datetime import datetime
-import os
 from pathlib import Path
-from tkinter import filedialog
 from tkinter import ttk
 import webbrowser
 
-from nvaeon2lib.json_timeline2 import JsonTimeline2
-from nvaeon2lib.nvaeon2_globals import _
+from nvaeon2.at2_service import At2Service
+from nvaeon2.nvaeon2_locale import _
 from nvlib.controller.plugin.plugin_base import PluginBase
-from nvlib.model.file.doc_open import open_document
-from nvlib.novx_globals import Error
-from nvlib.novx_globals import norm_path
 import tkinter as tk
 
 
@@ -38,33 +32,15 @@ class Plugin(PluginBase):
     API_VERSION = '5.0'
     DESCRIPTION = 'Synchronize with Aeon Timeline 2'
     URL = 'https://github.com/peter88213/nv_aeon2'
-    _HELP_URL = f'{_("https://peter88213.github.io/nvhelp-en")}/nv_aeon2/'
+    HELP_URL = f'{_("https://peter88213.github.io/nvhelp-en")}/nv_aeon2/'
 
     FEATURE = 'Aeon Timeline 2'
-    INI_FILENAME = 'nv_aeon2.ini'
-    INI_FILEPATH = '.novx/config'
-    SETTINGS = dict(
-        narrative_arc='Narrative',
-        property_description='Description',
-        property_notes='Notes',
-        property_moonphase='Moon phase',
-        type_arc='Arc',
-        type_character='Character',
-        type_location='Location',
-        type_item='Item',
-        role_arc='Arc',
-        role_plotline='Storyline',
-        role_character='Participant',
-        role_item='Item',
-        role_location='Location',
-        color_section='Red',
-        color_event='Yellow',
 
-    )
-    OPTIONS = dict(
-        add_moonphase=False,
-        lock_on_export=False,
-    )
+    def add_moonphase(self):
+        self.timelineService.add_moonphase()
+
+    def create_novx(self):
+        self.timelineService.create_novx()
 
     def disable_menu(self):
         """Disable menu entries when no project is open.
@@ -82,6 +58,15 @@ class Plugin(PluginBase):
         self._ui.toolsMenu.entryconfig(self.FEATURE, state='normal')
         self._timelineButton.config(state='normal')
 
+    def export_from_novx(self):
+        self.timelineService.export_from_novx()
+
+    def import_to_novx(self):
+        self.timelineService.import_to_novx()
+
+    def info(self):
+        self.timelineService.info()
+
     def install(self, model, view, controller):
         """Add a submenu to the main menu.
         
@@ -98,87 +83,48 @@ class Plugin(PluginBase):
         super().install(model, view, controller)
 
         # Create a submenu in the Tools menu.
-        self._pluginMenu = tk.Menu(self._ui.toolsMenu, tearoff=0)
-        self._ui.toolsMenu.add_cascade(label=self.FEATURE, menu=self._pluginMenu)
+        self.pluginMenu = tk.Menu(self._ui.toolsMenu, tearoff=0)
+        self._ui.toolsMenu.add_cascade(label=self.FEATURE, menu=self.pluginMenu)
         self._ui.toolsMenu.entryconfig(self.FEATURE, state='disabled')
-        self._pluginMenu.add_command(label=_('Information'), command=self._info)
-        self._pluginMenu.add_separator()
-        # self._pluginMenu.add_command(label=_('Settings'), command=self._edit_settings)
-        # self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Update the timeline'), command=self._export_from_novx)
-        self._pluginMenu.add_command(label=_('Update the project'), command=self._import_to_novx)
-        self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Add or update moon phase data'), command=self._add_moonphase)
-        self._pluginMenu.add_separator()
-        self._pluginMenu.add_command(label=_('Open Aeon Timeline 2'), command=self._launch_application)
+        self.pluginMenu.add_command(label=_('Information'), command=self.info)
+        self.pluginMenu.add_separator()
+        self.pluginMenu.add_command(label=_('Update the timeline'), command=self.export_from_novx)
+        self.pluginMenu.add_command(label=_('Update the project'), command=self.import_to_novx)
+        self.pluginMenu.add_separator()
+        self.pluginMenu.add_command(label=_('Add or update moon phase data'), command=self.add_moonphase)
+        self.pluginMenu.add_separator()
+        self.pluginMenu.add_command(label=_('Open Aeon Timeline 2'), command=self.launch_application)
 
         # Add an entry to the "File > New" menu.
-        self._ui.newMenu.add_command(label=_('Create from Aeon Timeline 2...'), command=self._create_novx)
+        self._ui.newMenu.add_command(label=_('Create from Aeon Timeline 2...'), command=self.create_novx)
 
         # Add an entry to the Help menu.
-        self._ui.helpMenu.add_command(label=_('Aeon 2 plugin Online help'), command=lambda: webbrowser.open(self._HELP_URL))
+        self._ui.helpMenu.add_command(label=_('Aeon 2 plugin Online help'), command=self.open_help_page)
 
         #--- Configure the toolbar.
         self._configure_toolbar()
+
+        self.timelineService = At2Service(model, view, controller, self.FEATURE)
+
+    def launch_application(self):
+        self.timelineService.launch_application()
 
     def lock(self):
         """Inhibit changes on the model.
         
         Overrides the superclass method.
         """
-        self._pluginMenu.entryconfig(_('Update the project'), state='disabled')
+        self.pluginMenu.entryconfig(_('Update the project'), state='disabled')
+
+    def open_help_page(self):
+        webbrowser.open(self.HELP_URL)
 
     def unlock(self):
         """Enable changes on the model.
         
         Overrides the superclass method.
         """
-        self._pluginMenu.entryconfig(_('Update the project'), state='normal')
-
-    def _add_moonphase(self):
-        """Add/update moon phase data.
-        
-        Add the moon phase to the event properties.
-        If the moon phase event property already exists, just update.
-        """
-        #--- Try to get persistent configuration data
-        if not self._mdl.prjFile:
-            return
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{JsonTimeline2.EXTENSION}'
-        if not os.path.isfile(timelinePath):
-            return
-
-        sourceDir = os.path.dirname(timelinePath)
-        if not sourceDir:
-            sourceDir = '.'
-        try:
-            homeDir = str(Path.home()).replace('\\', '/')
-            pluginCnfDir = f'{homeDir}/{self.INI_FILEPATH}'
-        except:
-            pluginCnfDir = '.'
-        iniFiles = [f'{pluginCnfDir}/{self.INI_FILENAME}', f'{sourceDir}/{self.INI_FILENAME}']
-        configuration = self._mdl.nvService.new_configuration(
-            settings=self.SETTINGS,
-            options=self.OPTIONS
-            )
-        for iniFile in iniFiles:
-            configuration.read(iniFile)
-        kwargs = {}
-        kwargs.update(configuration.settings)
-        kwargs.update(configuration.options)
-        kwargs['add_moonphase'] = True
-        kwargs['nv_service'] = self._mdl.nvService
-        timeline = JsonTimeline2(timelinePath, **kwargs)
-        timeline.novel = self._mdl.nvService.new_novel()
-        try:
-            timeline.read()
-            timeline.write(timeline.novel)
-        except Error as ex:
-            message = f'!{str(ex)}'
-        else:
-            message = f'{_("File written")}: "{norm_path(timeline.filePath)}".'
-        self._ui.set_status(message)
+        self.pluginMenu.entryconfig(_('Update the project'), state='normal')
 
     def _configure_toolbar(self):
 
@@ -206,7 +152,7 @@ class Plugin(PluginBase):
             self._ui.toolbar.buttonBar,
             text=_('Open Aeon Timeline 2'),
             image=aeon2Icon,
-            command=self._launch_application
+            command=self.launch_application
             )
         self._timelineButton.pack(side='left')
         self._timelineButton.image = aeon2Icon
@@ -221,186 +167,4 @@ class Plugin(PluginBase):
             return
 
         Hovertip(self._timelineButton, self._timelineButton['text'])
-
-    def _create_novx(self):
-        """Create a novelibre project from a timeline."""
-        timelinePath = filedialog.askopenfilename(
-            filetypes=[(JsonTimeline2.DESCRIPTION, JsonTimeline2.EXTENSION)],
-            defaultextension=JsonTimeline2.EXTENSION,
-            )
-        if not timelinePath:
-            return
-
-        self._ctrl.close_project()
-        root, __ = os.path.splitext(timelinePath)
-        novxPath = f'{root}{self._mdl.nvService.get_novx_file_extension()}'
-        kwargs = self._get_configuration(timelinePath)
-        kwargs['nv_service'] = self._mdl.nvService
-        source = JsonTimeline2(timelinePath, **kwargs)
-        target = self._mdl.nvService.new_novx_file(novxPath)
-
-        if os.path.isfile(target.filePath):
-            self._ui.set_status(f'!{_("File already exists")}: "{norm_path(target.filePath)}".')
-            return
-
-        message = ''
-        try:
-            source.novel = self._mdl.nvService.new_novel()
-            source.read()
-            target.novel = source.novel
-            target.write()
-        except Error as ex:
-            message = f'!{str(ex)}'
-        else:
-            message = f'{_("File written")}: "{norm_path(target.filePath)}".'
-            self._ctrl.open_project(filePath=target.filePath, doNotSave=True)
-        finally:
-            self._ui.set_status(message)
-
-    def _edit_settings(self):
-        """Toplevel window"""
-        return
-
-    def _export_from_novx(self):
-        """Update the timeline file from the novx file.
-        
-        Note:
-        The model's novel is not used as the conversion source here.
-        It is considered safer to use a copy of the novel read from the file.
-        """
-        if not self._mdl.prjFile:
-            return
-
-        self._ui.propertiesView.apply_changes()
-        self._ui.restore_status()
-        if not self._mdl.prjFile.filePath:
-            if not self._ctrl.save_project():
-                return
-                # cannot create a timeline if no novx project exists
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{JsonTimeline2.EXTENSION}'
-        if not os.path.isfile(timelinePath):
-            self._ui.set_status(_('!No {} file available for this project.').format(self.FEATURE))
-            return
-
-        if self._mdl.isModified:
-            if not self._ui.ask_yes_no(_('Save the project and update the timeline?')):
-                return
-
-            self._ctrl.save_project()
-        elif not self._ui.ask_yes_no(_('Update the timeline?')):
-            return
-
-        kwargs = self._get_configuration(timelinePath)
-        kwargs['nv_service'] = self._mdl.nvService
-        source = self._mdl.nvService.new_novx_file(self._mdl.prjFile.filePath, **kwargs)
-        source.novel = self._mdl.nvService.new_novel()
-        target = JsonTimeline2(timelinePath, **kwargs)
-        target.novel = self._mdl.nvService.new_novel()
-        try:
-            source.read()
-            target.read()
-            target.write(source.novel)
-            message = f'{_("File written")}: "{norm_path(target.filePath)}".'
-        except Error as ex:
-            message = f'!{str(ex)}'
-        self._ui.set_status(message)
-
-    def _get_configuration(self, sourcePath):
-        """ Read persistent configuration data for Aeon 2 conversion.
-        
-        First, look for a global configuration file in the novelibre installation directory,
-        then look for a local configuration file in the project directory.
-        """
-        sourceDir = os.path.dirname(sourcePath)
-        if not sourceDir:
-            sourceDir = '.'
-        try:
-            homeDir = str(Path.home()).replace('\\', '/')
-            pluginCnfDir = f'{homeDir}/{self.INI_FILEPATH}'
-        except:
-            pluginCnfDir = '.'
-        iniFiles = [f'{pluginCnfDir}/{self.INI_FILENAME}', f'{sourceDir}/{self.INI_FILENAME}']
-        configuration = self._mdl.nvService.new_configuration(
-            settings=self.SETTINGS,
-            options=self.OPTIONS
-            )
-        for iniFile in iniFiles:
-            configuration.read(iniFile)
-        kwargs = {}
-        kwargs.update(configuration.settings)
-        kwargs.update(configuration.options)
-        return kwargs
-
-    def _import_to_novx(self):
-        """Update the current project file from the timeline file.
-        
-        Note:
-        The NvWorkFile object of the open project cannot be used as target object.
-        This is because the JsonTimeline2 source object's IDs do not match, so 
-        the sections and other elements are identified by their titles when merging.
-        If anything goes wrong during the conversion, the model remains untouched.
-        """
-        if not self._mdl.prjFile:
-            return
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{JsonTimeline2.EXTENSION}'
-        if not os.path.isfile(timelinePath):
-            self._ui.set_status(_('!No {} file available for this project.').format(self.FEATURE))
-            return
-
-        if not self._ui.ask_yes_no(_('Save the project and update it?')):
-            return
-
-        self._ctrl.save_project()
-        kwargs = self._get_configuration(timelinePath)
-        kwargs['nv_service'] = self._mdl.nvService
-        source = JsonTimeline2(timelinePath, **kwargs)
-        target = self._mdl.nvService.new_novx_file(self._mdl.prjFile.filePath, **kwargs)
-        try:
-            target.novel = self._mdl.nvService.new_novel()
-            target.read()
-            source.novel = target.novel
-            source.read()
-            target.novel = source.novel
-            target.write()
-            message = f'{_("File written")}: "{norm_path(target.filePath)}".'
-            self._ctrl.open_project(filePath=self._mdl.prjFile.filePath, doNotSave=True)
-        except Error as ex:
-            message = f'!{str(ex)}'
-        self._ui.set_status(f'{message}')
-
-    def _info(self):
-        """Show information about the Aeon Timeline 2 file."""
-        if not self._mdl.prjFile:
-            return
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{JsonTimeline2.EXTENSION}'
-        if os.path.isfile(timelinePath):
-            try:
-                timestamp = os.path.getmtime(timelinePath)
-                if timestamp > self._mdl.prjFile.timestamp:
-                    cmp = _('newer')
-                else:
-                    cmp = _('older')
-                fileDate = datetime.fromtimestamp(timestamp).strftime('%c')
-                message = _('{0} file is {1} than the novelibre project.\n (last saved on {2})').format(self.FEATURE, cmp, fileDate)
-            except:
-                message = _('Cannot determine file date.')
-        else:
-            message = _('No {} file available for this project.').format(self.FEATURE)
-        self._ui.show_info(message, title=self.FEATURE)
-
-    def _launch_application(self):
-        """Launch Aeon Timeline 2 with the current project."""
-        if not self._mdl.prjFile:
-            return
-
-        timelinePath = f'{os.path.splitext(self._mdl.prjFile.filePath)[0]}{JsonTimeline2.EXTENSION}'
-        if os.path.isfile(timelinePath):
-            if self.OPTIONS['lock_on_export']:
-                self._ctrl.lock()
-            open_document(timelinePath)
-        else:
-            self._ui.set_status(_('!No {} file available for this project.').format(self.FEATURE))